+++ src/main/java/br/com/juliomartins/VassCommerce/model/Categoria.java
@@ -1,22 +1,22 @@
 package br.com.juliomartins.VassCommerce.model;
 
 public class Categoria {
-    private int id;
+    private long id;
     private String imagemSimboloUrl;
     private String nome;
     private String descricao;
 
-    public Categoria(int id, String imagemSimboloUrl, String nome, String descricao) {
+    public Categoria(long id, String imagemSimboloUrl, String nome, String descricao) {
         this.id = id;
         this.imagemSimboloUrl = imagemSimboloUrl;
         this.nome = nome;
         this.descricao = descricao;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
