+++ src/main/java/br/com/juliomartins/VassCommerce/model/Usuario.java
@@ -3,7 +3,7 @@ package br.com.juliomartins.VassCommerce.model;
 import java.util.Date;
 
 public class Usuario {
-    private int id;
+    private long id;
     private String nomeCompleto;
     private String  email;
     private String senha;
@@ -11,7 +11,7 @@ public class Usuario {
     private Date dataCadastro;
     private String fotoUrl;
 
-    public Usuario(int id, String nomeCompleto, String email, String senha,
+    public Usuario(long id, String nomeCompleto, String email, String senha,
                    Date dataCadastro, Date dataUltimaAtualizacao, String fotoUrl) {
         this.id = id;
         this.nomeCompleto = nomeCompleto;
@@ -22,11 +22,11 @@ public class Usuario {
         this.fotoUrl = fotoUrl;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
